Update object.service.ts
Added a check to see if the object had a record in the mall_object table when its dropped. If it doesn't, it will add that record. This allows objects added to CTR prior to this update to be added to the mall_objects table when placed in the shop.