Update StaffPage.vue
Added offset dropdown and pagination to warehouse page.
Added ability to destock an item from the mall and refund for unsold quantities. 
Added ability to set an object to deleted status.
Added error if the limit set is not a positive whole number.
Added ability to set an object back to Unlimited.
Fixed pagination.
Fixed spacing issues when displaying the objects on the lists.